refactor: move Codeblock to new file codeblock.py
